fix: 의존성 탭 동시 설치/업데이트 동시성 버그 수정 (#42)
의존성 탭에서 여러 항목의 설치/업데이트 버튼을 거의 동시에 누르면
바이너리가 손상되거나 UI 상태가 깨지던 문제를 프론트·백엔드 양쪽에서 수정.

백엔드(동시성)
- 의존성별 비동기 락 추가(dep_download::lock_dependency): 같은 의존성의
  install/update/delete를 직렬화. 기존엔 install_dependency·install_all
  ·update_dependency(별칭)·delete_app_managed_dep가 락 없이 동시에 돌며
  고정 temp 파일명(yt-dlp.tmp 등)과 최종 바이너리 경로를 truncate로 덮어써
  다운로드/추출이 손상될 수 있었음. 서로 다른 의존성은 경로가 달라 그대로 병렬.
- finalize_binary를 atomic rename 우선으로 변경: Unix에선 단일 rename으로
  교체(바이너리가 사라지는 구간 제거, 실행 중 프로세스는 기존 inode 유지),
  Windows에선 remove 후 rename로 폴백.

프론트(의존성 탭 상태)
- updating/installing을 단일 string에서 의존성별 Record로 변경: 두 번째
  작업이 시작돼도 첫 버튼이 다시 활성화되거나 스피너가 사라지지 않게 하고,
  같은 의존성의 중복 클릭을 막음.
- loadDepStatus에 in-flight 카운터 추가: 마지막 호출이 끝날 때만 스피너 종료.
- depActionResult를 의존성별로 분리해 동시 작업 결과가 서로 덮어쓰지 않게 함.

추가 비중복 수정
- history/logs 페이지네이션 offset의 u32 곱셈 오버플로를 u64 widening으로
  수정(db/queue.rs와 동일 패턴).
- 큐 페이지 취소 핸들러가 Result 상태를 확인하도록 형제 핸들러와 일치시킴.

검증: cargo fmt/clippy(-D warnings, 0건)/test(25 passed), bun run check(0 errors).

+++ src-tauri/src/ytdlp/db/history.rs
@@ -56,7 +56,8 @@ impl Database {
                 .map_err(|e| AppError::DatabaseError(e.to_string()))?
         };
 
-        let offset = page * page_size;
+        // Widen before multiplying so a large page can't overflow u32 (matches db/queue.rs).
+        let offset = (page as u64) * (page_size as u64);
         let query = format!(
             "SELECT id, video_url, video_id, title, quality_label, format, file_path, file_size, downloaded_at
              FROM history
